Avoid deprecation warnings when VOLK >= 3.1
Bump CMake max version to 3.28
Do not leak build system paths when cross-compiling. Fixes OpenEmbedded QA warning
volk_gnsssdr: Fix 64-bit integer testing
CI: Fix for Python 3.12

+++ src/algorithms/libs/volk_gnsssdr_module/volk_gnsssdr/lib/qa_utils.cc
@@ -483,12 +483,12 @@ bool icompare(t *in1, t *in2, unsigned int vlen, unsigned int tol)
     int print_max_errs = 10;
     for (unsigned int i = 0; i < vlen; i++)
         {
-            if (((unsigned int)abs(int(((t *)(in1))[i]) - int(((t *)(in2))[i]))) > tol)
+            if (((uint64_t)abs(int64_t(((t *)(in1))[i]) - int64_t(((t *)(in2))[i]))) > tol)
                 {
                     fail = true;
                     if (print_max_errs-- > 0)
                         {
-                            std::cout << "offset " << i << " in1: " << static_cast<int>(t(((t *)(in1))[i])) << " in2: " << static_cast<int>(t(((t *)(in2))[i]));
+                            std::cout << "offset " << i << " in1: " << static_cast<int64_t>(t(((t *)(in1))[i])) << " in2: " << static_cast<int64_t>(t(((t *)(in2))[i]));
                             std::cout << " tolerance was: " << tol << '\n';
                         }
                 }
